made dean's requested changes + fixed some styling imports
