+++ pageStyles/emailSender.styles.ts
@@ -1,15 +1,9 @@
-import { Divider, Table, TableRow, Typography } from '@mui/material'
+import { Table, TableRow } from '@mui/material'
 import { styled } from '@mui/material/styles'
-import { theme } from '../styles/theme'
 
 const SectionContainer = styled('div')({
   marginTop: 30,
   marginBottom: 30
-})
-
-const StyledDivider = styled(Divider)({
-  marginTop: 50,
-  marginBottom: 50
 })
 
 const StyledTable = styled(Table)({
@@ -26,15 +20,11 @@ const StyledTableContainer = styled('div')({
   marginBottom: 50
 })
 
-const StyledErrorMessage = styled(Typography)({
-  color: theme.palette.Red.main
-})
-
 const StyledFinalMessageContent = styled('div')({
   whiteSpace: 'pre-wrap'
 })
 
 export {
-  SectionContainer, StyledTableContainer, StyledDivider,
-  StyledTable, StyledTableRow, StyledFinalMessageContent, StyledErrorMessage
+  SectionContainer, StyledTableContainer,
+  StyledTable, StyledTableRow, StyledFinalMessageContent
 }
